Move retry mechanism to YfData._make_request()
Retry logic now applies to ALL API calls (download, history, info, earnings, etc)
instead of just download() function. This provides consistent retry behavior
across the entire library.

- Move _is_transient_error() from multi.py to data.py
- Implement retry loop in YfData._make_request() with exponential backoff
- Simplify _download_one() - remove duplicate retry logic
- Reduce net code by 57 lines (46 added, 103 deleted)
